% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpPanel.R
\name{helpPanel}
\alias{helpPanel}
\title{helpPanel}
\usage{
helpPanel(filePath, center = FALSE)
}
\arguments{
\item{filePath}{path to resource}

\item{center}{logical; whether to center or not.}
}
\value{
tagList
}
\description{
Creates a panel to insert a vignette in a centered way.
}
\author{
Reinhard Simon
}
