% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tutorials.R
\name{list_tutorials}
\alias{list_tutorials}
\title{list_tutorials}
\usage{
list_tutorials(pkg = NULL, name = "tutorial", lang = "en",
  format = ".html")
}
\arguments{
\item{pkg}{a vector of package names.}

\item{name}{character base file name}

\item{lang}{two characters for a language}

\item{format}{a file format; html (default) or other well know (.pdf; .docx)}
}
\value{
vector of path names
}
\description{
lists tutorials prepared as vignettes.
}
\details{
The idea is to scan packages for consistently named vignettes. This includes
different language versions. The naming scheme used here is:

[name]_[lang].html

where lang should be a two letter code.
}
\author{
Reinhard Simon
}

